requirements.txt

typing==3.10.0.2
python-dotenv==0.19.1
fastapi==0.74.0
pydantic==1.9.0
